Moved base and derived classes to separate files
