+++ src/optimization/design_parameterization/CMakeLists.txt
@@ -1,5 +1,5 @@
 set(DESIGN_PARAMETERIZATION_SRC
-    design_parameterization_base.cpp
+    design_parameterization_ffd.cpp
     )
 
 foreach(dim RANGE 1 3)
