+++ src/testing/euler_bump_optimization.cpp
@@ -29,6 +29,7 @@
 #include "optimization/rol_to_dealii_vector.hpp"
 #include "optimization/flow_constraints.hpp"
 #include "optimization/rol_objective.hpp"
+#include "optimization/design_parameterization/design_parameterization_ffd.hpp"
 
 #include "optimization/full_space_step.hpp"
