+++ tests/unit_tests/optimization/objective_check.cpp
@@ -25,6 +25,7 @@
 #include "optimization/rol_to_dealii_vector.hpp"
 #include "optimization/flow_constraints.hpp"
 #include "optimization/rol_objective.hpp"
+#include "optimization/design_parameterization/design_parameterization_ffd.hpp"
 
 
 const double TOL = 1e-7;
